requirements.txt
scikit-image
scikit-learn
numpy
scipy
sounddevice
tensorflow==1.13.1
Keras==2.2.4
librosa
llvmlite
matplotlib
Pillow